Add dashboard demo to the showcase
