+++ examples/showcase/Cargo.toml
@@ -8,3 +8,4 @@ yew = { path = "../.." }
 counter = { path = "sub/counter" }
 crm = { path = "sub/crm" }
 custom_components = { path = "sub/custom_components" }
+dashboard = { path = "sub/dashboard" }
