+++ examples/showcase/sub/dashboard/Cargo.toml
@@ -7,4 +7,4 @@ authors = ["Denis Kolodin <deniskolodin@gmail.com>"]
 failure = "0.1"
 serde = "1"
 serde_derive = "1"
-yew = { path = "../../.." }
+yew = { path = "../../../.." }
